+++ examples/vue2/vite.config.ts
@@ -1,13 +1,26 @@
 import { defineConfig } from 'vite'
-import { createVuePlugin } from 'vite-plugin-vue2'
+import vue from '@vitejs/plugin-vue2'
 import { resolve } from 'path'
 
 export default defineConfig({
-	plugins: [createVuePlugin()],
+	plugins: [vue()],
 	resolve: {
 		alias: {
 			'@': resolve(__dirname, 'src'),
-			vue: 'vue/dist/vue.esm.js',
 		},
+	},
+	build: {
+		rollupOptions: {
+			external: [],
+			output: {
+				globals: {},
+			},
+		},
+		commonjsOptions: {
+			transformMixedEsModules: true,
+		},
+	},
+	define: {
+		'process.env.NODE_ENV': JSON.stringify(process.env.NODE_ENV || 'development'),
 	},
 })